--- packages/itwin/tree-widget/src/components/core-trees/category-tree/CategoriesTree.scss
@@ -1,9 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-@import "../VisibilityTreeBase";
-
-.ui-fw-categories-tree {
-  @include ui-fw-visibility-tree-base;
-}
