--- packages/itwin/tree-widget/src/components/core-trees/index.ts
@@ -1,13 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-
-export * from "./Common";
-export * from "./VisibilityTreeEventHandler";
-export * from "./VisibilityTreeRenderer";
-export * from "./category-tree/CategoriesTree";
-export * from "./category-tree/CategoryVisibilityHandler";
-export * from "./models-tree/ModelsTree";
-export * from "./models-tree/ModelsVisibilityHandler";
-export * from "./spatial-tree/SpatialContainmentTree";
