--- packages/itwin/tree-widget/src/components/trees/ModelsTree.scss
@@ -1,14 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-
-@import "~@itwin/core-react/lib/cjs/core-react/index";
-
-.tree-widget-models-tree-container {
-  flex: 1;
-}
-
-.tree-widget-models-tree-toolbar-icon {
-  color: $buic-icon-color;
-}
