+++ packages/itwin/tree-widget/src/components/trees/VisibilityTreeBase.scss
