+++ packages/itwin/tree-widget/src/components/trees/category-tree/Categories.json
