+++ packages/itwin/tree-widget/src/components/trees/index.ts
@@ -1,9 +1,17 @@
 /*---------------------------------------------------------------------------------------------
- * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
- * See LICENSE.md in the project root for license terms and full copyright notice.
- *--------------------------------------------------------------------------------------------*/
+* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+* See LICENSE.md in the project root for license terms and full copyright notice.
+*--------------------------------------------------------------------------------------------*/
 
-export * from "./ModelsTree";
-export * from "./CategoriesTree";
-export * from "./SpatialTree";
+export * from "./Common";
 export * from "./IModelContentTree";
+export * from "./VisibilityTreeEventHandler";
+export * from "./VisibilityTreeRenderer";
+export * from "./category-tree/CategoriesTree";
+export * from "./category-tree/CategoriesTreeComponent";
+export * from "./category-tree/CategoryVisibilityHandler";
+export * from "./models-tree/ModelsTree";
+export * from "./models-tree/ModelsTreeComponent";
+export * from "./models-tree/ModelsVisibilityHandler";
+export * from "./spatial-tree/SpatialContainmentTree";
+export * from "./spatial-tree/SpatialTreeComponent";
