+++ packages/itwin/tree-widget/src/components/trees/models-tree/Hierarchy.GroupedByClass.json
