+++ packages/itwin/tree-widget/src/components/trees/models-tree/ModelsTree.scss
@@ -3,11 +3,16 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 @import "~@itwin/core-react/lib/cjs/core-react/index";
+@import "../VisibilityTreeBase";
 
-.tree-widget-category-tree-search-bar {
-  margin-bottom: $uicore-s;
+.ui-fw-models-tree {
+  @include ui-fw-visibility-tree-base;
 }
 
-.tree-widget-category-tree-toolbar-icon {
+.tree-widget-models-tree-container {
+  flex: 1;
+}
+
+.tree-widget-models-tree-toolbar-icon {
   color: $buic-icon-color;
 }
