+++ packages/itwin/tree-widget/src/components/trees/models-tree/ModelsTreeComponent.tsx
@@ -6,16 +6,16 @@
 import React, { useCallback, useEffect, useState } from "react";
 import type { Viewport } from "@itwin/core-frontend";
 import { useActiveIModelConnection, useActiveViewport } from "@itwin/appui-react";
-import { useTreeFilteringState } from "../TreeFilteringState";
+import { useTreeFilteringState } from "../../TreeFilteringState";
 import "./ModelsTree.scss";
 import type {
   GeometricModel3dProps,
   ModelQueryParams,
 } from "@itwin/core-common";
-import { TreeHeaderComponent } from "../header/TreeHeader";
-import type { ModelTreeProps } from "../../types";
-import { AutoSizer } from "../utils/AutoSizer";
-import { ModelsTree } from "../core-trees/models-tree/ModelsTree";
+import { TreeHeaderComponent } from "../../header/TreeHeader";
+import type { ModelTreeProps } from "../../../types";
+import { AutoSizer } from "../../utils/AutoSizer";
+import { ModelsTree } from "./ModelsTree";
 
 interface TreeViewModelInfo {
   id: string;
