+++ packages/itwin/tree-widget/src/components/trees/spatial-tree/SpatialBreakdown.GroupedByClass.json
