+++ packages/itwin/tree-widget/src/components/trees/spatial-tree/SpatialContainmentTree.scss
