+++ packages/itwin/tree-widget/src/test/trees/models-tree/ModelsVisibilityHandler.test.ts
@@ -12,8 +12,8 @@ import { QueryRowFormat } from "@itwin/core-common";
 import { IModelApp, NoRenderApp, PerModelCategoryVisibility } from "@itwin/core-frontend";
 import { isPromiseLike } from "@itwin/core-react";
 import { Presentation } from "@itwin/presentation-frontend";
-import { ModelsVisibilityHandler } from "../../../components/core-trees/models-tree/ModelsVisibilityHandler";
-import { CachingElementIdsContainer } from "../../../components/core-trees/models-tree/Utils";
+import { ModelsVisibilityHandler } from "../../../components/trees/models-tree/ModelsVisibilityHandler";
+import { CachingElementIdsContainer } from "../../../components/trees/models-tree/Utils";
 import { TestUtils } from "../../TestUtils";
 import { createCategoryNode, createElementClassGroupingNode, createElementNode, createModelNode, createSubjectNode } from "../Common";
 import type { Id64String } from "@itwin/core-bentley";
@@ -21,7 +21,7 @@ import type {
   IModelConnection, SpatialViewState, Viewport, ViewState, ViewState3d,
 } from "@itwin/core-frontend";
 import type { FilteredPresentationTreeDataProvider } from "@itwin/presentation-components";
-import type { ModelsVisibilityHandlerProps } from "../../../components/core-trees/models-tree/ModelsVisibilityHandler";
+import type { ModelsVisibilityHandlerProps } from "../../../components/trees/models-tree/ModelsVisibilityHandler";
 import type { IModelHierarchyChangeEventArgs, PresentationManager } from "@itwin/presentation-frontend";
 
 describe("ModelsVisibilityHandler", () => {
