+++ packages/itwin/tree-widget/src/tree-widget-react.ts
@@ -7,6 +7,5 @@ export * from "./types";
 export * from "./TreeWidget";
 export * from "./components/TreeWidgetComponent";
 export * from "./components/header/TreeHeader";
-export * from "./components/core-trees";
 export * from "./components/trees";
 export * from "./components/TreeWidgetUiItemsProvider";
